% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_dockerfile.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{Create dockerfiles for a given R version}
\usage{
create_dockerfile(r_version)
}
\arguments{
\item{r_version}{A vector of strings.}
}
\value{
Called for it's side effect (returns the first argument invisibly).
}
\description{
Create dockerfiles for a given R version
}
\examples{
\dontrun{
r_version <- c("3.6.0", "3.5.0")
create_dockerfile(r_version)
}
}
