% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_dockerfile.R
\name{enlist_dockerfile}
\alias{enlist_dockerfile}
\title{Create a list of docker files from a vector or r versions}
\usage{
enlist_dockerfile(r_version)
}
\arguments{
\item{r_version}{A vector of strings.}
}
\value{
A list of strings.
}
\description{
Create a list of docker files from a vector or r versions
}
\examples{
r_version <- glue::glue("3.{6:2}.0")
enlist_dockerfile(r_version)
}
