% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_dockerfile.R
\name{enlist_dockerfiles}
\alias{enlist_dockerfiles}
\title{Create a list of docker files from a vector or r versions}
\usage{
enlist_dockerfiles(r_version)
}
\arguments{
\item{r_version}{A vector of strings.}
}
\value{
A list of strings.
}
\description{
Create a list of docker files from a vector or r versions
}
\examples{
r_version <- c("3.6.0", "3.5.0")
docker_files <- enlist_dockerfiles(r_version)
docker_files
}
